update solidity to 0.8.16

+++ solidity/libsolidity/analysis/ConstantEvaluator.cpp
@@ -27,6 +27,8 @@
 #include <libsolidity/ast/TypeProvider.h>
 #include <liblangutil/ErrorReporter.h>
 
+#include <limits>
+
 using namespace std;
 using namespace solidity;
 using namespace solidity::frontend;
@@ -47,7 +49,7 @@ bool fitsPrecisionExp(bigint const& _base, bigint const& _exp)
 
 	size_t const bitsMax = 4096;
 
-	unsigned mostSignificantBaseBit = (unsigned)boost::multiprecision::msb(_base);
+	size_t mostSignificantBaseBit = static_cast<size_t>(boost::multiprecision::msb(_base));
 	if (mostSignificantBaseBit == 0) // _base == 1
 		return true;
 	if (mostSignificantBaseBit > bitsMax) // _base >= 2 ^ 4096
